gemini-live-agent/backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.14.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
asgiref==3.10.0
asttokens==3.0.0
asyncio-throttle==1.0.2
asyncpg==0.30.0
attrs==25.3.0
audioop-lts==0.2.2
Authlib==1.6.9
backoff==2.2.1
banks==2.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==25.9.0
blinker==1.9.0
blis==1.3.0
build==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chroma-hnswlib==0.7.6
chromadb==1.1.1
click==8.2.1
cloudpathlib==0.22.0
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.11
confection==0.1.5
crewai==0.165.1
cryptography==46.0.5
cymem==2.0.11
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docx2txt==0.9
durationpy==0.10
email-validator==2.3.0
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.12.0
fastapi==0.135.1
filelock==3.19.1
filetype==1.2.0
Flask==3.1.3
flatbuffers==25.2.10
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
gkeepapi==0.15.1
google-adk==1.27.1
google-ai-generativelanguage==0.9.0
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.49.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.141.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.36.2
google-cloud-bigtable==2.35.0
google-cloud-core==2.4.3
google-cloud-dataplex==2.16.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.21.0
google-cloud-logging==3.14.0
google-cloud-monitoring==2.29.1
google-cloud-pubsub==2.26.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.63.0
google-cloud-speech==2.37.0
google-cloud-storage==2.18.2
google-cloud-trace==1.18.0
google-crc32c==1.7.1
google-genai==1.67.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gpsoauth==2.0.0
graphviz==0.21
greenlet==3.2.4
griffe==1.14.0
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.75.0
grpcio-status==1.62.3
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
instagrapi==2.0.0
instructor==1.10.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain-core==0.3.81
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.4.14
language_data==1.3.0
litellm==1.74.9
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.10.28
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-google==0.4.1
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-instrumentation==0.4.2
llama-index-legacy==0.9.48.post4
llama-index-llms-gemini==0.2.0
llama-index-llms-google-genai==0.6.2
llama-index-llms-openai==0.1.31
llama-index-multi-modal-llms-gemini==0.6.1
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-google==0.3.1
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-postgres==0.7.1
llama-index-workflows==2.8.3
llama-parse==0.4.9
llamaindex-py-client==0.1.19
loguru==0.7.3
lxml==6.0.0
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==2.1.1
neo4j==5.28.2
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.0
oauth2client==4.1.3
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.22.0
openai==1.109.1
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
orjson==3.11.2
ormsgpack==1.12.1
overrides==7.7.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pgvector==0.4.1
pi_heif==1.1.0
pikepdf==10.1.0
pillow==11.0.0
platformdirs==4.5.0
playwright==1.58.0
portalocker==2.7.0
posthog==5.4.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.6
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybase64==1.4.2
pycparser==2.22
pycryptodomex==3.18.0
pydantic==2.12.5
pydantic-settings==2.6.1
pydantic_core==2.41.5
pydeck==0.9.1
PyDrive==1.3.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
pyOpenSSL==26.0.0
pyparsing==3.2.5
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2025.2.18
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.22
python-oxmsg==0.0.2
pytokens==0.3.0
pyttsx3==2.99
pytz==2025.2
pyvis==0.3.2
pywin32==311
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==3.3.1
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.36
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
srsly==2.5.1
sse-starlette==3.3.2
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.52.1
streamlit==1.48.1
striprtf==0.0.26
sympy==1.14.0
tavily-python==0.7.17
tenacity==9.1.4
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
unstructured==0.18.13
unstructured-client==0.42.3
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.12
uvicorn==0.41.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.6
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.17.3
xlrd==2.0.1
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
